First upload
Geo coordinate retrieval based on user-entered location is working
These coordinates are then fed to the sunrise-sunset API, and the sunrise and sunset times
for the location are sent back, and displayed to the user

TO FIX:
multiple times are being output. the previous entry is not being cleared out.